Convert DampingMoment listener to simulation extension
