+++ core/src/net/sf/openrocket/simulation/extension/example/DampingMomentProvider.java
@@ -1,0 +1,13 @@
+package net.sf.openrocket.simulation.extension.impl;
+
+import net.sf.openrocket.plugin.Plugin;
+import net.sf.openrocket.simulation.extension.AbstractSimulationExtensionProvider;
+
+@Plugin
+public class DampingMomentProvider extends AbstractSimulationExtensionProvider {
+	
+	public DampingMomentProvider() {
+		super(DampingMoment.class, "Post-step flight conditions", "Damping Moment Coefficient (Cdm)");
+	}
+	
+}
